+++ organize/cli.py
@@ -5,52 +5,84 @@ The file management automation tool.
 """
 import os
 import sys
-from pathlib import Path
-from typing import Tuple, Optional
+import textwrap
+from typing import Optional, Tuple
 
 import click
-from fs import appfs
+import fs
 
 from . import console
 from .__version__ import __version__
 from .migration import NeedsMigrationError
 
-# Alternative: "https://tfeldmann.github.io/organize/"
-DOCS_URL = "https://organize.readthedocs.io"
-MIGRATE_URL = "https://organize.readthedocs.io/en/latest/updating-from-v1/"
-DEFAULT_CONFIG = """\
-# organize configuration file
-# {docs}
+DOCS_RTD = "https://organize.readthedocs.io"
+DOCS_GHPAGES = "https://tfeldmann.github.io/organize/"
 
-rules:
-  - locations:
-      - # your locations here
-    filters:
-      - # your filters here
-    actions:
-      - # your actions here
-""".format(
-    docs=DOCS_URL
-)
+DEFAULT_CONFIG_FS_URL = "userconf://organize::/config.yaml"
 
-try:
-    config_filename = "config.yaml"
-    envvar = os.getenv("ORGANIZE_CONFIG")
-    if envvar:
-        dirname, config_filename = os.path.split(envvar)
-        if not dirname or not config_filename:
-            raise ValueError("Invalid ORGANIZE_CONFIG: %s" % envvar)
-        CONFIG_PATH = Path(envvar)
-    else:
-        CONFIG_PATH = Path(appfs.UserConfigFS("organize").getsyspath(config_filename))
 
-    # create default config file if it not exists
-    CONFIG_PATH.parent.mkdir(parents=True, exist_ok=True)
-    if not CONFIG_PATH.exists():
-        CONFIG_PATH.write_text(DEFAULT_CONFIG)
-except Exception as e:
-    console.error(str(e), title="Config file")
-    sys.exit(1)
+def path_split(path_or_url: str) -> Tuple[str, str]:
+    if sys.platform.startswith("win"):
+        path_or_url = path_or_url.replace("\\", "/")
+    dirname, filename = fs.path.split(path_or_url)
+    return dirname, filename
+
+
+def ensure_default_config():
+    """
+    Ensures a configuration file exists in the default location.
+    """
+    DEFAULT_CONFIG_TEXT = textwrap.dedent(
+        """\
+        # organize configuration file
+        # {docs}
+
+        rules:
+          - name: "The name of this rule"
+            locations:
+              - # your locations here
+            filters:
+              - # your filters here
+            actions:
+              - # your actions here
+        """
+    ).format(docs=DOCS_RTD)
+
+    dirname, filename = path_split(DEFAULT_CONFIG_FS_URL)
+    if not filename:
+        raise ValueError("invalid config path, missing filename")
+    with fs.open_fs(dirname, create=True, writeable=True) as confdir:
+        if not confdir.exists(filename):
+            confdir.writetext(filename, DEFAULT_CONFIG_TEXT)
+
+
+def read_config(fs_url: str) -> Tuple[str, str]:
+    """
+    Read the config at the given fs_url.
+    """
+    dirname, filename = path_split(fs_url)
+    with fs.open_fs(dirname) as confdir:
+        try:
+            config_path = confdir.getsyspath(filename)
+        except fs.errors.NoSysPath:
+            config_path = fs_url
+        return (config_path, confdir.readtext(filename))
+
+
+def read_default_config() -> Tuple[str, str]:
+    """
+    Read the config file set in $ORGANIZE_CONFIG and fall back to the default
+    config folder if this env is not set.
+    """
+    # first check whether the user set a env var
+    env_fs_url = os.getenv("ORGANIZE_CONFIG")
+
+    # if no env variable is given we make sure that there is a config file in the
+    # default location
+    if not env_fs_url:
+        ensure_default_config()
+
+    return read_config(env_fs_url or DEFAULT_CONFIG_FS_URL)
 
 
 class NaturalOrderGroup(click.Group):
@@ -70,53 +102,60 @@ class TagType(click.ParamType):
 CLI_CONFIG = click.argument(
     "config",
     required=False,
-    default=CONFIG_PATH,
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
+    type=str,
 )
 CLI_WORKING_DIR_OPTION = click.option(
     "--working-dir",
-    default=".",
-    type=click.Path(exists=True, file_okay=False, dir_okay=True, path_type=Path),
+    default="",
+    type=str,
     help="The working directory",
 )
-# for CLI backwards compatibility with organize v1.x
-CLI_CONFIG_FILE_OPTION = click.option(
-    "--config-file",
-    default=None,
-    hidden=True,
-    type=click.Path(exists=True, dir_okay=False, path_type=Path),
+CLI_TAGS = click.option(
+    "--tags",
+    type=TagType(),
+    default="",
+    help="tags to run",
 )
-CLI_TAGS = click.option("--tags", type=TagType(), default="")
-CLI_SKIP_TAGS = click.option("--skip-tags", type=TagType(), default="")
+CLI_SKIP_TAGS = click.option(
+    "--skip-tags",
+    type=TagType(),
+    default="",
+    help="tags to skip",
+)
 
 
-def run_local(
-    config_path: Path,
-    working_dir: Path,
+def execute(
+    config: Optional[str],
+    working_dir: str,
     simulate: bool,
     tags: Optional[Tuple[str]] = None,
     skip_tags: Optional[Tuple[str]] = None,
 ):
     from schema import SchemaError
-
     from . import core
 
+    if config:
+        config_path, config_text = read_config(config)
+    else:
+        config_path, config_text = read_default_config()
+
     try:
-        console.info(config_path=config_path, working_dir=working_dir)
-        config = config_path.read_text()
-        os.chdir(working_dir)
+        console.info(config=config_path, working_dir=working_dir)
         core.run(
-            rules=config,
+            rules=config_text,
             simulate=simulate,
+            working_dir=working_dir,
             tags=tags,
             skip_tags=skip_tags,
         )
     except NeedsMigrationError as e:
+        from .migration import MIGRATION_DOCS_URL
+
         console.error(e, title="Config needs migration")
         console.warn(
             "Your config file needs some updates to work with organize v2.\n"
             "Please see the migration guide at\n\n"
-            "%s" % MIGRATE_URL
+            "%s" % MIGRATION_DOCS_URL
         )
         sys.exit(1)
     except SchemaError as e:
@@ -144,18 +183,12 @@ def cli():
 @cli.command()
 @CLI_CONFIG
 @CLI_WORKING_DIR_OPTION
-@CLI_CONFIG_FILE_OPTION
 @CLI_TAGS
 @CLI_SKIP_TAGS
-def run(config: Path, working_dir: Path, config_file, tags, skip_tags):
+def run(config: Optional[str], working_dir: str, tags, skip_tags):
     """Organizes your files according to your rules."""
-    if config_file:
-        config = config_file
-        console.deprecated(
-            "The --config-file option can now be omitted. See organize --help."
-        )
-    run_local(
-        config_path=config,
+    execute(
+        config=config,
         working_dir=working_dir,
         simulate=False,
         tags=tags,
@@ -166,18 +199,12 @@ def run(config: Path, working_dir: Path, config_file, tags, skip_tags):
 @cli.command()
 @CLI_CONFIG
 @CLI_WORKING_DIR_OPTION
-@CLI_CONFIG_FILE_OPTION
 @CLI_TAGS
 @CLI_SKIP_TAGS
-def sim(config: Path, working_dir: Path, config_file, tags, skip_tags):
+def sim(config: Optional[str], working_dir: str, tags, skip_tags):
     """Simulates a run (does not touch your files)."""
-    if config_file:
-        config = config_file
-        console.deprecated(
-            "The --config-file option can now be omitted. See organize --help."
-        )
-    run_local(
-        config_path=config,
+    execute(
+        config=config,
         working_dir=working_dir,
         simulate=True,
         tags=tags,
@@ -186,12 +213,7 @@ def sim(config: Path, working_dir: Path, config_file, tags, skip_tags):
 
 
 @cli.command()
-@click.argument(
-    "config",
-    required=False,
-    default=CONFIG_PATH,
-    type=click.Path(),
-)
+@click.argument("config", required=False, type=str)
 @click.option(
     "--editor",
     envvar="EDITOR",
@@ -208,7 +230,7 @@ def edit(config, editor):
 @cli.command()
 @CLI_CONFIG
 @click.option("--debug", is_flag=True, help="Verbose output")
-def check(config: Path, debug):
+def check(config: str, debug):
     """Checks whether a given config file is valid.
 
     If called without arguments it will check the default config file.
@@ -218,9 +240,12 @@ def check(config: Path, debug):
     from .core import highlighted_console as out
     from .core import replace_with_instances
 
-    print("Checking: %s" % config)
     try:
-        config_str = config.read_text()
+        if config:
+            config_path, config_str = read_config(config)
+        else:
+            config_path, config_str = read_default_config()
+        print("Checking: %s" % config_path)
 
         if debug:
             out.rule("Raw", align="left")
@@ -258,11 +283,11 @@ def check(config: Path, debug):
             out.print("\n\n")
             out.rule("Instantiation", align="left")
 
-        warnings = replace_with_instances(rules)
-        if debug:
-            out.print(rules)
-            for msg in warnings:
-                out.print("Warning: %s" % msg)
+        # warnings = replace_with_instances(rules)
+        # if debug:
+        #     out.print(rules)
+        #     for msg in warnings:
+        #         out.print("Warning: %s" % msg)
 
         if debug:
             out.print("\n\n")
@@ -278,10 +303,29 @@ def check(config: Path, debug):
 @click.option("--path", is_flag=True, help="Print the path instead of revealing it.")
 def reveal(path: bool):
     """Reveals the default config file."""
+    fs_url = os.getenv("ORGANIZE_CONFIG", DEFAULT_CONFIG_FS_URL)
+
+    dirname, filename = fs.path.split(fs_url)
+    try:
+        with fs.open_fs(dirname) as dirfs:
+            syspath = dirfs.getsyspath(filename)
+    except Exception:
+        syspath = None
+
     if path:
-        click.echo(CONFIG_PATH)
+        if syspath:
+            click.echo(syspath)
+        else:
+            click.echo(fs_url)
+        return
+
+    if syspath:
+        import webbrowser
+
+        webbrowser.open("file://%s" % fs.path.dirname(syspath))
     else:
-        click.launch(str(CONFIG_PATH), locate=True)
+        click.echo("Revealing this fs_url is not possible.")
+        click.echo(fs_url)
 
 
 @cli.command()
@@ -303,7 +347,9 @@ def schema():
 @cli.command()
 def docs():
     """Opens the documentation."""
-    click.launch(DOCS_URL)
+    import webbrowser
+
+    webbrowser.open(DOCS_RTD)
 
 
 # deprecated - only here for backwards compatibility