+++ tests/conftest.py
@@ -1,10 +1,33 @@
+from typing import Union
 from unittest.mock import patch
 
+import fs
 import pytest
 from fs.base import FS
 from fs.path import basename, join
 
-from organize import config
+
+@pytest.fixture
+def tempfs():
+    with fs.open_fs("temp://") as tmp:
+        yield tmp
+
+
+@pytest.fixture
+def memfs():
+    with fs.open_fs("mem://") as mem:
+        yield mem
+
+
+@pytest.fixture(
+    params=[
+        "mem://",
+        # pytest.param("temp://", marks=pytest.mark.skip),
+    ],
+)
+def testfs(request) -> FS:
+    with fs.open_fs(request.param) as tmp:
+        yield tmp
 
 
 @pytest.fixture
@@ -13,7 +36,7 @@ def mock_echo():
         yield mck
 
 
-def make_files(fs: FS, layout: dict, path="/"):
+def make_files(fs: FS, layout: Union[dict, list], path="/"):
     """
     layout = {
         "folder": {
@@ -26,6 +49,10 @@ def make_files(fs: FS, layout: dict, path="/"):
     }
     """
     fs.makedirs(path, recreate=True)
+    if isinstance(layout, list):
+        for f in layout:
+            fs.touch(f)
+        return
     for k, v in layout.items():
         respath = join(path, k)
 
@@ -51,31 +78,3 @@ def read_files(fs: FS, path="/"):
     for x in fs.walk.dirs(path, max_depth=0):
         result[basename(x)] = read_files(fs, path=join(path, x))
     return result
-
-
-def rules_shortcut(fs: FS, filters, actions, location="files", max_depth=0):
-    if isinstance(filters, str):
-        filters = config.load_from_string(filters)
-    if isinstance(actions, str):
-        actions = config.load_from_string(actions)
-
-    # for action in actions:
-    #     for opts in action.values():
-    #         if "filesystem" in opts and opts["filesystem"] == "mem":
-    #             opts["filesystem"] = fs
-
-    return {
-        "rules": [
-            {
-                "locations": [
-                    {
-                        "path": location,
-                        "filesystem": fs,
-                        "max_depth": max_depth,
-                    }
-                ],
-                "actions": actions,
-                "filters": filters,
-            }
-        ]
-    }